move login tab tests to their own test file
